requirements.txt
Flask
pymongo
flask_cors
python-dotenv